backend/requirements.txt
fastapi
uvicorn
numpy
opencv-python
python-multipart
sqlalchemy
ultralytics
python-jose[cryptography]
passlib[bcrypt]
